pyproject.toml
[build-system]
requires = ["setuptools>=70.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brenda-agentic-workflow"
version = "0.1.0"
description = "Production-ready multi-agentic workflow focused on BRENDA enzymatic data mining and analysis."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Viku" }
]
license = { text = "Proprietary" }
dependencies = [
  "langchain>=0.2.0",
  "langchain-community>=0.2.0",
  "langchain-openai>=0.1.7",
  "pydantic>=2.6.0",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.1",
  "httpx>=0.27.0",
  "tenacity>=8.2.3",
  "openai>=1.30.0",
  "pandas>=2.2.2",
  "sqlalchemy>=2.0.29",
  "requests>=2.31.0",
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "redis>=5.0.0",
  "structlog>=24.1.0",
  "aiofiles>=23.2.1",
  "ijson>=3.2.3",
  "tabulate>=0.9.0",
  "rich>=13.7.1",
  "gradio>=4.44.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.5",
  "pytest-cov>=5.0.0",
  "mypy>=1.10.0",
  "ruff>=0.4.4",
  "black>=24.4.2",
  "pre-commit>=3.7.0",
  "ipykernel>=6.29.4",
  "notebook>=7.1.2",
  "rich>=13.7.1"
]

[tool.black]
line-length = 100
target-version = ["py310"]
include = "src"

[tool.ruff]
line-length = 100
src = ["src"]
select = ["E", "F", "B", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-ra --cov=src --cov-report=term-missing"
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_unreachable = true
strict = false
plugins = ["pydantic.mypy"]
check_untyped_defs = true
ignore_missing_imports = true
